refactor: HomeView 수정 (#3)
- CampaignSection -> CampainBanner로 명칭 수정
