+++ cherrydan/Presentation/Home/HomeView.swift
@@ -2,32 +2,39 @@ import SwiftUI
 
 struct HomeView: View {
     @State private var selectedTab = 0
+    
     let buttonArr = ["인기 공고", "마감임박", "새로운 공고", "선정확률 높은", "관심 공고", "지역 맞춤"]
     
+    let campaigns = [
+        ["체험단 캠페인 여기 다 모았다!","지금 최고의 체험단에 신청해 보세요.", "1"],
+        ["체험단 캠페인 여기 다 모았다!","지금 최고의 체험단에 신청해 보세요.", "2"],
+        ["체험단 캠페인 여기 다 모았다!","지금 최고의 체험단에 신청해 보세요.", "3"]
+    ]
+    
     var body: some View {
-        ScrollView {
-            VStack(spacing: 20) {
-                header
-                    .padding(.horizontal, 16)
-                
-                CampaignSection()
-                
-                ScrollView(.horizontal, showsIndicators: false) {
-                    HStack(spacing: 0) {
-                        ForEach(0..<buttonArr.count, id: \.self) { index in
-                            categorySection(index)
-                                .onTapGesture { selectedTab = index }
-                        }
-                        .padding(.horizontal, 8)
-                    }
-                    .padding(.horizontal, 8)
+        VStack(spacing: 0) {
+            CDHeaderWithLeftContent() {
+                HStack(alignment: .center) {
+                    Text("카테고리")
+                        .font(.t1)
                     
+                    Image("hamburger")
                 }
-                
-                campaignGridSection
+            }
+            
+            ScrollView {
+                VStack(spacing: 20) {
+                    CampaignBanner(campaigns: campaigns)
+                    
+                    CDTabSection(
+                        selectedTab: $selectedTab,
+                        data: buttonArr
+                    )
+                    
+                    campaignGridSection
+                }
             }
         }
-        .animation(.spring, value: selectedTab)
     }
     
     private var campaignGridSection: some View {
@@ -50,44 +57,8 @@ struct HomeView: View {
         }
         .padding(.horizontal)
     }
-    
-    private func categorySection(_ index:Int) -> some View {
-        VStack(spacing: 8){
-            Text(buttonArr[index])
-                .font(.m3b)
-                .foregroundColor(selectedTab == index ? .mPink3 : .gray4)
-                .padding(.horizontal, 4)
-            
-            Rectangle()
-                .fill(index == selectedTab ? .mPink3 : .clear)
-                .frame(height: 2)
-        }
-    }
 }
 
-private var header: some View {
-    HStack(alignment: .center) {
-        Text("카테고리")
-            .font(.t1)
-        Image("hamburger")
-        
-        Spacer()
-        
-        HStack(spacing: 4) {
-            Button(action: {}) {
-                Image("alarm")
-            }
-            
-            Button(action: {}) {
-                Image("search")
-            }
-        }
-    }
-    .padding(.top, 8)
+#Preview {
+    HomeView()
 }
-
-struct HomeView_Previews: PreviewProvider {
-    static var previews: some View {
-        HomeView()
-    }
-}
